Guest Instance Playbook Revamp
The guest.instances.yml playbook is simplified.

First of all, the majority of the shell code is moved to a single
file; shell scripts includes that file and invoke the function
there. It limits most code generation to that single file. Other
template files are also simplified as well. As a result, template
file 'guest.instances.inc.j2' is be removed entirely.

Also, to simplify configuration and make the playbook more robust,
the systemd service file installed removes the memory lock limit
entirely, instead of guessing how much the limit should be lifted.