Explicit Set the MAC Address of Bridge Interface
Currently, the "bridge" network interface does not have an explicit
MAC address assigned to it. A side effect of this is that the bridge
may select the smallest MAC address of all the enslaved interface and
use it as its own MAC address. As enslaved interfaces come and go, the
the MAC address of the bridge interface can also drift. The drifting
MAC address can pose problem for IP address assignment.

To fix the problem, we explicitly set the bridge MAC address to be
that of a permanently enslaved ethernet card.