Fcitx Related Cleanup for Firejailed Apps
No fcitx hacks are needed for firejailed X applications when the
profile do not use either "x11" or "machine-id" options. Fcitx
resolves the private dbus connection in a two-step process. The
first step is to locate the file where the details of the private
connection is stored. It can be done by deriving the path from
Machine ID file and X display number. It can also be overridden
using the FCITX_DBUS_ADDRESS environment variable. The second
step is to parse the file, check the existence of fcitx/dbus
processes and connect to the abstract UNIX socket specified by
the file.