Sandbox for Apache Service
The systemd unit of apache service is updated to sandbox the server
from the rest of the system. User directories are hidden from the
server; system directories except those used by the server are made
available to the server for reading only. Lastly, many irrelevant
syscalls are forbidden by the server process.