Sanitize Main Apache Configuration
The main apache configuration is sanitized as follow. First of all,
the local customization files are now prefixed with zzz- instead of
local_ to ensure that they are loaded last and overrides whatever
in the default files. Next, the files are now managed using
debian/ubuntu tools. Finally, the default CGI configuration is
disabled.