Update Dpkg Architecture Support Differently
This commit changes how i386 architecture is enabled in dpkg.

Before the change, dpkg is executed on the remote side to make the
change. The command is idempotent so running the command multiple
times is safe. However, Ansible has no way to determine if the
command makes an effective change to the system and therefore always
reports the execution as a change.