Major Update to Desktop Software Install Playbook
The desktop software install playbook receives a major update and
refactor as follow:

Firefox installation is moved to a separate playbook. Debian only
provides package for Firefox ESR but not the standard version.
Therefore, the playbook installs the standard version from Mozilla
itself.

Wine installation is also moved to a separate playbook. The playbook
can install either from distribution or official website. It also
installs Mono and Gecko in the shared directory.

Finally, package list are updated, and Ubuntu specific data is
removed from the playbook.