Major Update on Bumblebee Service
The bumblebee playbook receives a major update. As we move to Debian
and it carries i386 build of related libraries, the playbook can use
system packages directly. No compilation is needed and a number of
config changes are no longer required.

Moreover, the playbook also creates a script to setup the correct
PRIMUS_VK_RENDERID and PRIMUS_VK_DISPLAYID environment variables.