New Playbooks for Grub, Plymouth, SDDM, etc.
New playbooks are created to configure Grub, Plymouth, SDDM and
KScreenLocker. The new playbooks are created to make sure that the
packages uses the proper themes and are visually consistent.

In particular, Grub suffers a bug where background image stored in
encrypted root will not be available to Grub during boot, forcing
Grub to show a plain text boot menu rather than a graphical one.
The playbook fixes the bug.