Skype Firejail Profile Update
The default firejail profile for Skype does not work. The problem
is that Skype is installed into /usr/share/skypeforlinux, but that
location is not executable due to the AppArmor setting. The commit
fixes the problem by whitelisting the directory in the AppArmor
settings.

Also, unprivileged usernamespace is supported on Debian, so we enable
the relevant part in the generic Electron profile.