Revamped Baidu Net Disk Playbook
The Baidu Net Disk Playbook receives a major revamp.

Most importantly, the current official debian package is broken
and the client frequently segfaults right after launch. However,
if the electron runtime in the package is replaced with a later
electron release from upstream the problem is resolved. Therefore,
the installation procedure is changed from installing the official
package to cherrypicking files from electron and package to form
a working chimera.

In addition to that, a firejail profile for Baidu Net Disk is
installed for extra protection.