Improvement to Firejail Networking Support
The firejail playbook is updated to improve networking support.
First of all, the playbook uses dnsmasq to provide DNS service to
confined apps. Next, the configuration options is simplified to be
less repetitive. Then, extra include files are created for inclusion
in profiles to activate network isolation. Lastly, the netfilter
files are renamed to match the referencing include files.