Update Account Specific Zoom Settings
The commit makes some changes to account specific zoom settings.
First of all, the desktop file is now generated from the system
desktop file using the firefox_xdg_desktop_file module. Then, the
firejail profile is customized to enable access to some part of
the "Downloads" folder. Lastly, the firejail profile is customized
to allow internet access but not other systems on the local area
network.