Update Account Specific Baidu Net Disk Settings
The commit makes some changes to account specific baidu net disk
settings. First of all, the desktop file is now generated from
the system desktop file using the firefox_xdg_desktop_file module.
Then, the firejail profile is modified to rely on global profile.
Moreover, the firejail profile is customized to enable access to
some part of the "Downloads" folder. Finally, the firejail profile
is customized to allow internet access but not other systems in the
local area network.