Update Account Specific Steam Settings
The commit makes some changes to account specific steam settings.
First of all, the desktop file is now generated from the system
desktop file using the firefox_xdg_desktop_file module. Then,
the firejail profile is customized to rely on the global profile.
Last but not least, the environment variables are now set in the
firejail profile instead of command line.