Update Firejail Profile for Steam
The commit includes a few changes to firejail profile for Steam.

First of all, pulseaudio now puts the socket to /run/user/uid.
Therefore, we need to whitelist the file in the profile. We do so
by including the whitelist-runuser-common.inc.

Additionally, extra directories are whitelisted in the account
specific profile.