Update the System Setup Playbooks
The system setup playbooks are updated to make them suitable
for Fedora. The commit contains quite a lot of changes:

First of all, playbooks files now starts with system prefix
instead of software, which better reflect their roles in
setting up important system services.

Secondly, Debian related playbooks are replaced with Fedora
ones. For example, playbook for dpkg is replaced with one
for dnf.

Thirdly, a new playbook is added for setting up ramdrive
instead of manual setup.

Then, firejail configuration is updated. In particular,
network related code is temporarily removed due to disuse
and pending playbook for firewall.

Last but not least, network manager is also updated to not
deal with IP forwarding, also due to pending playbook for
firewall.