Update the Software Setup Playbooks
The software setup playbooks are updated to make the suitable
for Fedora. The commit contains quite a lot of changes:

First of all, the playbook now uses dnf, not dpkg, to install
packages for obvious reason.

Secondly, the software list are adjusted. For example, fcitx
is upgraded to version 5. Some new console utilities like
nvtop are added, etc.

Thirdly, some of the software are now installed via flatpak

Last but not least, playbook files now start with software
prefix instead of install.