Setup Libvirt and Windows Guest
The commit adds two playbooks, one for setting up libvirt and the
other setting up a windows guest.

The libvirt playbook installs and enables libvirt on the local
machine. It also enables polkit based ACL for libvirt.

The windows playbook creates a libvirt domain which is suitable
for installing windows. It also adds some scripts to control the
domain. At the moment, the looking glass client is compiled and
installed manually. It should be later automated in the playbook.