Revamp to Manage Multiple Machines
The playbooks now have to manage multiple machines. Since the
machines use different operating systems from Fedora, Alma Linux
to OpenWRT, it is easier to write playbooks for each machine
instead of adding machine specific checks to existing playbooks.

This commit renames the existing playbooks by adding the prefix
"desktop" to indicate that the playbooks are intended for the
desktop machine.