KVM Module Options Template Bugfix
The template files/etc/desktop/etc/modprobe.d/kvm.conf.j2 sets any
missing option to 'null'. But 'null' is not a valid jinja literal,
so when any option is not specified the template will fail due to
the invalid literal. Instead, the correct literal is 'None'. This
commit replaces the incorrect 'null' literal with the correct 'None'
literal.