Playbook for Zigbee2MQTT on the Server
This commit adds a playbook for installing and configuring zigbee2mqtt
on the server.

Note that the podman quadlet uses Volume instead of AddDevice to add
Zigbee adapters to the container. It is due to some interaction between
device node creation and user namespace in podman which causes the passed
devicefile to have incorrect permission.