Rename Home Assistant Quadlet and Firewalld Service
The quadlet file and firewalld service names of Home Assistant
is updated to replace underscore with hyphen.

For podman quadlets, generated container names and unit file
names use hyphen as separator. For firewalld services, all
builtin services use hyphen as separator as well. Therefore,
this change should improve the consistency between the service
and other parts of the system.